import updates of integration tests, renamed test_utils
