+++ python/tests/integration/completions/test_chat_completion_with_function_calling.py
@@ -25,7 +25,7 @@ from tests.integration.completions.chat_completion_test_base import (
     vertex_ai_setup,
 )
 from tests.integration.completions.completion_test_base import ServiceType
-from tests.integration.test_utils import retry
+from tests.integration.utils import retry
 
 if sys.version_info >= (3, 12):
     from typing import override  # pragma: no cover
